change how BUILDKITE_PLUGIN_AWS_PARAMSTORE_SECRETS_KEY works.
We used to allow user to set this value to point us to an envvar that we
use as the slug, however, in most cases we only care for per repo and
per pipeline settings. So we've made change in the bootstrap script to
set BUILDKITE_PLUGIN_AWS_PARAMSTORE_SECRETS_MODE to either repo or
pipeline, and we will hardcode the envvars in these two scenarios.

Also did some rename to match the flattened code structure.

Signed-off-by: Kai Xia <kaix+github@fastmail.com>